left/right join IFNA() added.
A separate lookup is used to find records where there is no match between tables.  If the case where NO records were found, it would return "NA" and cause an array error.  So a IFNA() test is added that returns an empty array so that when included in the JOIN array - no error is generated.